requirements.txt
streamlit
requests
pandas
numpy
ortools
fpdf2
